Make copy_slice operate on pointers instead of slices
This function previously took slices, where one slice pointed to
rust-allocated memory (fine), and one slices pointed to mmap'd guest
memory (potentially problematic). The latter could result in
undefined behavior, as we cannot assume exclusive ownership of
guest memory if a virtual machine is running: We need to assume
that the guest inside the VM implicitly holds a mutable reference
to the entire guest memory.

By using pointers instead of reference, this is still a problem
due to data races being possible, but the worst case scenario here
is inconsistent reads/writes, whereas the aliasing rules are
something the compiler potentially uses for optimizations.

+++ src/bytes.rs
@@ -120,7 +120,7 @@ pub unsafe trait ByteValued: Copy + Default + Send + Sync {
     /// be no other accesses).
     fn as_bytes(&mut self) -> VolatileSlice {
         // SAFETY: This is safe because the lifetime is the same as self
-        unsafe { VolatileSlice::new(self as *mut Self as usize as *mut _, size_of::<Self>()) }
+        unsafe { VolatileSlice::new(self as *mut Self as *mut _, size_of::<Self>()) }
     }
 }
